+++ ecs/maps_gen_test.go
@@ -31,6 +31,19 @@ func TestMap1(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap1NewBatch(t *testing.T) {
@@ -84,6 +97,29 @@ func TestMap1NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap1Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap1[ChildOf](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap2(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -107,6 +143,19 @@ func TestMap2(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap2NewBatch(t *testing.T) {
@@ -160,6 +209,29 @@ func TestMap2NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap2Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap2[ChildOf, CompB](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap3(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -183,6 +255,19 @@ func TestMap3(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{}, &CompC{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap3NewBatch(t *testing.T) {
@@ -236,6 +321,29 @@ func TestMap3NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap3Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap3[ChildOf, CompB, CompC](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, &CompC{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap4(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -259,6 +367,19 @@ func TestMap4(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{}, &CompC{}, &CompD{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap4NewBatch(t *testing.T) {
@@ -312,6 +433,29 @@ func TestMap4NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap4Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap4[ChildOf, CompB, CompC, CompD](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, &CompC{}, &CompD{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap5(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -335,6 +479,19 @@ func TestMap5(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{}, &CompC{}, &CompD{}, &CompE{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap5NewBatch(t *testing.T) {
@@ -388,6 +545,29 @@ func TestMap5NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap5Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap5[ChildOf, CompB, CompC, CompD, CompE](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap6(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -411,6 +591,19 @@ func TestMap6(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, &CompF{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap6NewBatch(t *testing.T) {
@@ -464,6 +657,29 @@ func TestMap6NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap6Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap6[ChildOf, CompB, CompC, CompD, CompE, CompF](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, &CompF{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap7(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -487,6 +703,19 @@ func TestMap7(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, &CompF{}, &CompG{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap7NewBatch(t *testing.T) {
@@ -540,6 +769,29 @@ func TestMap7NewBatchFn(t *testing.T) {
 	assert.Equal(t, 3*n, cnt)
 }
 
+func TestMap7Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap7[ChildOf, CompB, CompC, CompD, CompE, CompF, CompG](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, &CompF{}, &CompG{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
+}
+
 func TestMap8(t *testing.T) {
 	n := 12
 	w := NewWorld(4)
@@ -563,6 +815,19 @@ func TestMap8(t *testing.T) {
 	for _, e := range entities {
 		mapper.Remove(e)
 	}
+
+	assert.Panics(t, func() {
+		mapper.Get(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.HasAll(Entity{})
+	})
+	assert.Panics(t, func() {
+		mapper.Add(Entity{}, &CompA{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, &CompF{}, &CompG{}, &CompH{})
+	})
+	assert.Panics(t, func() {
+		mapper.Remove(Entity{})
+	})
 }
 
 func TestMap8NewBatch(t *testing.T) {
@@ -614,4 +879,27 @@ func TestMap8NewBatchFn(t *testing.T) {
 	}
 	assert.True(t, mapper.HasAll(lastEntity))
 	assert.Equal(t, 3*n, cnt)
+}
+
+func TestMap8Relations(t *testing.T) {
+	w := NewWorld(8)
+
+	mapper := NewMap8[ChildOf, CompB, CompC, CompD, CompE, CompF, CompG, CompH](&w)
+
+	parent1 := w.NewEntity()
+	parent2 := w.NewEntity()
+
+	e := mapper.NewEntity(&ChildOf{}, &CompB{}, &CompC{}, &CompD{}, &CompE{}, &CompF{}, &CompG{}, &CompH{}, Rel(0, parent1))
+	assert.Equal(t, parent1, mapper.GetRelation(e, 0))
+	assert.Equal(t, parent1, mapper.GetRelationUnchecked(e, 0))
+
+	mapper.SetRelations(e, Rel(0, parent2))
+	assert.Equal(t, parent2, mapper.GetRelation(e, 0))
+
+	assert.Panics(t, func() {
+		mapper.SetRelations(Entity{}, Rel(0, parent2))
+	})
+	assert.Panics(t, func() {
+		mapper.GetRelation(Entity{}, 0)
+	})
 }