+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Adds `ExchangeX.Add` and `ExchangeX.Remove` (#70)
 - Adds `Map.GetRelationUnchecked` (#75)
 - Adds build tag `tiny` for 64 bit masks (#77)
+- Adds `QueryX.GetRelation`, `MapX.GerRelation` and `MapX.GetRelationUnchecked` (#79)
 
 ## [[v0.1.0]](https://github.com/mlange-42/ark/tree/v0.1.0)